+++ src/app/Data/services/forecast-data/forecast-data.service.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map, Observable } from 'rxjs';
 
-import { environment } from '../../../environments/environment';
-import { ForecastData } from '../../Domain/models/forecast-data.model';
+import { environment } from '../../../../environments/environment';
+import { ForecastData } from '../../../Domain/models/forecast-data.model';
 
 @Injectable({ providedIn: 'root' })
 export class ForecastDataService {