+++ src/app/Domain/state/forecast-data/forecast-data.effect.ts
@@ -5,7 +5,7 @@ import { Store } from '@ngrx/store';
 
 import * as ForecastActions from './forecast-data.actions';
 import { AppState } from '../app.state';
-import { ForecastDataService } from '../../../Data/services/forecast-data.service';
+import { ForecastDataService } from '../../../Data/services/forecast-data/forecast-data.service';
 import { ForecastData } from '../../models/forecast-data.model';
 
 @Injectable({ providedIn: 'root' })