+++ CHANGELOG.md
@@ -59,3 +59,8 @@ All notable changes to this project will be documented in this file.
 
 - Modal component.
 - Modal component at pushing switch user.
+
+
+## [0.5.0]
+
+- Tests for the forecastService.